% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_matings.R
\name{arrange_matings}
\alias{arrange_matings}
\title{return list of  who mates with whom to create hybrid offspring.}
\usage{
arrange_matings(A, B)
}
\arguments{
\item{A}{a data frame with id, pop, and group.  Should be exclusively from one of the groups (i.e. wild)}

\item{B}{a data frame with id and pop, and group.  Should be from the other group (i.e. farmed)}
}
\value{
a data frame with columns id, popA, popB, idA gamA idB gamB that tells us which gametes from
which individuals each new individual is to be made of.
}
\description{
This will create F1's between
the A and the B group.
}
\keyword{internal}

