% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_train_test.R
\name{assign_train_test}
\alias{assign_train_test}
\title{assign individuals to training and test sets.}
\usage{
assign_train_test(dat, train_fract = 0.5)
}
\arguments{
\item{dat}{data frame with columns id, variant, gene_copy, allele,
chrom, coord, pop, and group.}

\item{train_fract}{what fraction of each pop should be training (the rest is test)}
}
\value{
A data frame just like dat, but with a column called "test_or_train" that
has entries of "test" or "train" as appropriate for each fish id.
}
\description{
Simple version that does a fixed fraction from each pop
}
\keyword{internal}
