% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epp.R
\name{compute_epp}
\alias{compute_epp}
\title{Compute the expected posterior probability of assignment of gene copies.}
\usage{
compute_epp(D)
}
\arguments{
\item{D}{a data frame like that returned by assign_train_test()}
}
\value{
a data frame that includes variant, chrom, coord, allele then has
two columns of allele dplyr::counts (farmed and wild) taken only from the "train"
individuals and then two columns of allele freqs (with the 1/2 correction
in there) and  epp computed from those allele freqs.
}
\description{
See the notebooks for an explanation.
}
\keyword{internal}
