% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hybrid_dataset.R
\name{create_hybrid_dataset}
\alias{create_hybrid_dataset}
\title{Create a simulated data set for newhybrids using dplyr::selected/ranked markers}
\usage{
create_hybrid_dataset(SAR, wild_pop, hyb_cat, L, dir = "hyb_dir",
  file_name = "nh_data.txt")
}
\arguments{
\item{SAR}{the output of split_and_rank()}

\item{wild_pop}{a vector of the names of the populations of wild fish to simulate from. It treats them
as all being from the same population if they are included together.  This is not a 
problem for F1's, but it is likely unrealistic for F2s and backcrosses. Unless the 
F1 hybrids have no site fidelity (i.e. they stray freely...)}

\item{hyb_cat}{the hybrid category desired}

\item{L}{the number of loci to choose}

\item{dir}{the path of the directory to create to put the result into.  It will 
write the result into a file called \code{file_name}.}

\item{file_name}{the name of the output file.  Defaults to "nh_data.txt".}
}
\description{
We do this so that all the training individuals will end up with `z0s` or `z1s` in the data
set, and then we make as many of the others as we can.  Since we are going to be just computing 
the scaled likelihoods for all these guys, it is fine to just create everyone to be of one type
(i.e. Pure, F1, F2, Bx, etc.)  This let's you make sure that backcrosses are all within the same
population.
}

