% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bxs.R
\name{make_bxs}
\alias{make_bxs}
\title{Make backcrosses by segregating chromosome segments}
\usage{
make_bxs(Wf, Ff, idstr = "BX")
}
\arguments{
\item{Wf}{a data frame of individuals with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}

\item{Ff}{a data frame of individuals (from the other species/population)
with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}

\item{idstr}{string that gets prepended to the number of each individual}
}
\description{
blab
}
\keyword{internal}

