% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_f1s.R
\name{make_f1s}
\alias{make_f1s}
\title{Mate two collections of individuals together to make as many offspring as possible}
\usage{
make_f1s(Wf, Ff)
}
\arguments{
\item{Wf}{a data frame of individuals with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}

\item{Ff}{a data frame of individuals (from the other species/population)
with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}
}
\description{
This does sampling without replacement from the genes of the individuals.
Basically, it takes individuals that have haplotypes in hap1 and hap2, and it
segregates two gametes for each individual (with recombination) and then those
get combined with gametes from other individuals.  If there are W indivs in
Wf and F indivs in Ff, this function will return 2 * min(W, F) offspring from mating between
min(W, F) randomly dplyr::selected individuals in Wf and Ff.
}
\keyword{internal}
