% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_f2s.R
\name{make_f2s}
\alias{make_f2s}
\title{make as many F2's as possible}
\usage{
make_f2s(Wf, Ff)
}
\arguments{
\item{Wf}{a data frame of individuals with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}

\item{Ff}{a data frame of individuals (from the other species/population)
with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}
}
\description{
internal function
}
\keyword{internal}
