% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pures.R
\name{make_pures}
\alias{make_pures}
\title{make pure individuals by segregating chunks of genome amongst founders}
\usage{
make_pures(Df)
}
\arguments{
\item{Df}{a data frame of individuals with haplotypes in hap1 and hap2. If these are
founders then the haplotypes should have been scrambled first, using
\code{\link{scramble_founder_haplotypes}}.}
}
\description{
Since founders are scrambled, this is not super meaningful, but it is still
a good way to simulate new genotypes.
internal function
}
\keyword{internal}

