% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{marker_selection}
\alias{marker_selection}
\title{select most informative markers on chromosomes}
\usage{
marker_selection(R, min_dist = 60000)
}
\arguments{
\item{R}{a data frame like that returned by compute_epp()}
}
\value{
returns a data frame with chrom, coord, variant, epp,
and a new column "selectable" which is TRUE if it is one that should
be selected, and FALSE if it is too close to another, previously selected
marker.  It comes back sorted, descending on epp.
}
\description{
Does not select markers that are within min_dist base pairs
of a previously selected marker.
}
\keyword{internal}
