% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh_tablify.R
\name{nh_tablify}
\alias{nh_tablify}
\title{turn the genotypes of individuals into rows for a new hybrids file}
\usage{
nh_tablify(D, V, opt_str = "", id_prepend = "")
}
\arguments{
\item{D}{a tidy data frame with at least the columns: id, variant, gene_copy, allele}

\item{V}{a vector of variant names giving which variants to use and the order they are
to be in.}

\item{opt_str}{The options string to be added to each individual (like "z0s")}

\item{id_prepend}{this string will be prepended to the fish id (i.e. use it for "train_", or "test_")}
}
\description{
This also makes both gene copies missing if only one of them was.  This happens in the
simulated hybrids, so obviously must be dealt with.
}
\keyword{internal}
