% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2tidy.R
\name{plink2tidy}
\alias{plink2tidy}
\title{Convert a plink file to a tidy data frame}
\usage{
plink2tidy(ped, map)
}
\arguments{
\item{ped}{path to a space delimited plink ped file}

\item{map}{path to a tab delimited plink map file}
}
\description{
This  might be useful to users.
}

