% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_selection.R
\name{ranked_selector}
\alias{ranked_selector}
\title{an internal function to be applied to each chromosome}
\usage{
ranked_selector(bp, md)
}
\arguments{
\item{md}{the min_dist allowed between a new SNP and any other selected ones}

\item{co}{a vector of base pair coordinates ranked descending on epp}
}
\description{
this is a very internal function.  Users should never see this.
}
\keyword{internal}

