% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_founder_haplotypes.R
\name{scramble_founder_haplotypes}
\alias{scramble_founder_haplotypes}
\title{Randomly assign (no-LD) markers to one chromosome or another in  founder individuals}
\usage{
scramble_founder_haplotypes(D, V)
}
\arguments{
\item{D}{a tidy data frame with at least the columns: id, variant, gene_copy, allele, chrom, coord, pop}

\item{V}{a vector of variant names giving which variants to use and the order they are
to be in.}
}
\value{
A data frame with all the identifiers as before, but now the alleles are in two
columns: hap1 and hap2, which are the alleles as if they occurred together on different
homologous chromosomes.  This gets sorted by id, chrom, and coord.
}
\description{
blah blah
}
\keyword{internal}

