% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregate_gametes.R
\name{segregate_gametes}
\alias{segregate_gametes}
\title{Segregate the markers on two haploypes on each individual into two gametes.}
\usage{
segregate_gametes(H, Mb_recomb_prob = 0.01)
}
\arguments{
\item{H}{a data frame like the one that comes out of scramble_founder_haplotypes().  It has
to have hap1 and hap2 columns.}

\item{Mb_recomb_prob}{the probability of recombination over 10^6 bp (one Mb).
By default this is set to 0.01 (i.e. a centi-Morgan).}
}
\description{
Used internally.
}
\keyword{internal}

