% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_rank.R
\name{split_and_rank}
\alias{split_and_rank}
\title{split data set into training and test and rank markers based on training set}
\usage{
split_and_rank(dat, train_fract = 0.5, min_dist = 60000)
}
\arguments{
\item{dat}{a data frame that has the columns: id, variant, gene_copy, allele,
chrom, coord, pop, group.}

\item{train_fract}{fraction of individuals from each group (and pop, too) to 
assign to the training set.}

\item{min_dist}{the number of base pairs below which you will not select a marker if
another one within min_dist on the same chromosome has already been selected.}
}
\value{
a list with components `split_dat` and `ranked_markers`.
}
\description{
We are going to have a single function that splits the same into training and test 
and also ranks the loci.
}

