% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nh.R
\name{write_nh}
\alias{write_nh}
\title{write a new_hybs file}
\usage{
write_nh(G, path = "nh_data.txt")
}
\arguments{
\item{G}{a data frame like that prepared by nh_tablify, or perhaps several of those
that have been dplyr::bind_rows()-ed together.}

\item{path}{the file path you want to write to.}
}
\description{
Takes care of the preamble and everthing.
}
\keyword{internal}

